+++ jeap-messaging-sequential-inbox-test/src/test/java/ch/admin/bit/jeap/messaging/sequentialinbox/integrationtest/SequentialInboxWithEncryptionIT.java
@@ -3,7 +3,9 @@ package ch.admin.bit.jeap.messaging.sequentialinbox.integrationtest;
 import ch.admin.bit.jeap.crypto.api.KeyId;
 import ch.admin.bit.jeap.crypto.api.KeyIdCryptoService;
 import ch.admin.bit.jeap.messaging.sequentialinbox.integrationtest.encryption.CryptoServiceTestConfig;
+import ch.admin.bit.jeap.messaging.sequentialinbox.persistence.SequenceInstanceState;
 import ch.admin.bit.jme.declaration.JmeDeclarationCreatedEvent;
+import ch.admin.bit.jme.test.JmeSimpleTestEvent;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
@@ -14,11 +16,11 @@ import org.springframework.boot.test.autoconfigure.actuate.observability.AutoCon
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.TestPropertySource;
 
 import java.util.UUID;
 
-import static ch.admin.bit.jeap.messaging.sequentialinbox.integrationtest.message.TestMessages.createDeclarationCreatedEvent;
-import static ch.admin.bit.jeap.messaging.sequentialinbox.integrationtest.message.TestMessages.randomContextId;
+import static ch.admin.bit.jeap.messaging.sequentialinbox.integrationtest.message.TestMessages.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.eq;
 
@@ -28,6 +30,7 @@ import static org.mockito.ArgumentMatchers.eq;
         classes = {TestApp.class, CryptoServiceTestConfig.class}
 )
 @Slf4j
+@TestPropertySource(properties = "jeap.messaging.sequential-inbox.config-location=classpath:/messaging/jeap-sequential-inbox-for-encryption-and-signature.yml")
 @ActiveProfiles({"message-encryption-enabled", "key-id-crypto-service"})
 class SequentialInboxWithEncryptionIT extends SequentialInboxITBase {
 
@@ -44,22 +47,24 @@ class SequentialInboxWithEncryptionIT extends SequentialInboxITBase {
     void encryptMessage_and_receiveEncryptedMessage() {
 
         final KeyId testKeyId = KeyId.of("testKey");
-
-        // given: a test event
         UUID contextId = randomContextId();
-        JmeDeclarationCreatedEvent event = createDeclarationCreatedEvent(contextId);
 
-        // when: sending the event
+        JmeDeclarationCreatedEvent event = createDeclarationCreatedEvent(contextId);
         sendSync(JmeDeclarationCreatedEvent.TypeRef.DEFAULT_TOPIC, event);
 
-        // then: assert that the event was consumed by the message listener
+        assertSequenceState(contextId.toString(), SequenceInstanceState.OPEN);
+
+        JmeSimpleTestEvent jmeSimpleTestEvent = createJmeSimpleTestEvent(contextId);
+        sendSync(JmeSimpleTestEvent.TypeRef.DEFAULT_TOPIC, jmeSimpleTestEvent);
+
         assertMessageConsumedByListener(event);
         assertSequencedMessageProcessedSuccessfully(event);
-        assertSequenceOpen(event);
-        assertBufferedMessageCount(contextId, 0);
+        assertSequenceState(contextId.toString(), SequenceInstanceState.CLOSED);
+        assertBufferedMessageCount(contextId, 1);
+        assertSequenceClosed(contextId);
 
         Mockito.verify(keyIdCryptoService).encrypt(plainMessageCaptor.capture(), eq(testKeyId));
-        Mockito.verify(keyIdCryptoService).decrypt(encryptedMessageCaptor.capture());
+        Mockito.verify(keyIdCryptoService, Mockito.times(2)).decrypt(encryptedMessageCaptor.capture());
         byte[] plainMessage = plainMessageCaptor.getValue();
         byte[] encryptedMessage = encryptedMessageCaptor.getValue();
         assertThat(plainMessage).isNotEqualTo(encryptedMessage);